requirements.txt
discord.py>=2.4,<3
firebase-admin>=6.6,<7
python-dotenv>=1.0,<2
Flask>=3.1,<4
httpx>=0.28,<0.29
gunicorn>=23,<24
